Interactable crosshair V1
Made it so all things that are interactable make the crosshair change as a indication! 2 new scripts. 1 called iinteractable its just a reference script so ui script and crosshair highlitghter can take reference from it so no need to add it anywhere just keep it in project folder scripts. crosshair highlighter is added onto player object as a component and uses a reference from ui script that is located in the canvas crosshair